requirements.txt
torch==2.0.1
torchmetrics==1.4.2
torchvision==0.15.2
accelerate==1.0.0
addict==2.4.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
appdirs==1.4.4
args==0.1.0
asttokens==2.4.1
attrs==24.2.0
av==13.1.0
black==24.1.0
blessed==1.20.0
braceexpand==0.1.7
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clint==0.5.1
clip==0.2.0
clip-benchmark==1.6.1
cmake==3.30.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
crcmod==1.7
cryptography==43.0.1
cycler==0.12.1
dataclasses==0.6
DataProperty==1.0.1
datasets==2.16.1
debugpy==1.8.5
decorator==5.1.1
decord==0.6.0
deepspeed==0.14.4
diffusers==0.31.0
dill==0.3.7
distlib==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
evaluate==0.4.3
executing==2.1.0
ExifRead-nocycle==3.0.1
fairscale==0.4.13
filelock==3.14.0
fire==0.5.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
h11==0.14.0
hf_transfer==0.1.8
hjson==3.1.0
httpcore==0.16.3
httpx==0.23.3
identify==2.6.1
idna==3.8
image-reward==1.5
imageio==2.36.0
img2dataset==1.45.0
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==0.10.0
joblib==1.4.2
jsonlines==4.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
libcst==1.4.0
lightning-utilities==0.11.7
lit==18.1.8
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mbstrdecoder==1.1.3
mdurl==0.1.2
mmcv==2.0.0
mmcv-full==1.7.2
mmengine==0.7.3
model-index==0.1.11
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
mutagen==1.47.0
mypy==1.5.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
nodeenv==1.9.1
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open_clip_torch==2.26.1
openai==1.51.2
openai-clip==1.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.1.1
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.1
peft==0.13.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.10.1
pre_commit==4.0.1
prompt_toolkit==3.0.47
protobuf==3.20.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pybind11==2.13.6
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
pyparsing==3.1.4
pytablewriter==1.2.0
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2023.4
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.7.24
requests==2.32.3
rfc3986==1.5.0
rich==13.4.2
ruff==0.6.9
sacrebleu==2.4.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.1.99
sentry-sdk==2.13.0
setproctitle==1.3.3
shapely==2.0.6
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlitedict==2.1.0
stack-data==0.6.3
sympy==1.13.2
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.6
tenacity==8.3.0
termcolor==2.5.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
timm==0.6.13
tokenizers==0.19.1
tomli==2.0.2
tornado==6.4.1
tqdm==4.65.2
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers-stream-generator==0.0.5
triton==2.0.0
typepy==1.3.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
virtualenv==20.26.6
wandb==0.16.6
wcwidth==0.2.13
webdataset==0.2.100
websockets==13.1
xxhash==3.5.0
yapf==0.40.2
yarl==1.13.1
yt-dlp==2024.10.7
zipp==3.20.2
zss==1.2.0
zstandard==0.23.0
flash_attn==2.6.3